refactor: move the SYNC loading and its state to app-sync.js, the version badge to the app shell (#33)
app-sync.js owns the loaded statuses, the loading and failure flags, the load
button handler, the painting of the badges and the SYNC controls. It never
imports app.js: the selected project and SYNC filter are read through the
accessors given to initializeSyncControls, and onLoadEnd lets app.js re-apply
its filters after a load (and keep the SYNC filter at Show all while no
statuses are loaded, as before). fetchAndDisplayVersion is banner chrome and
moves to app-shell.js, called by initializeAppShell.
